pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-forwarder"
version = "1.0.5"
description = "Telegram forwarding dashboard with multi-rule filtering, account/Bot delivery, history search, and Docker support."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Telethon>=1.42.0,<2.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "python-socks[asyncio]>=2.5.3,<3.0.0",
  "PySocks>=1.7.1,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.32.0,<1.0.0",
  "python-multipart>=0.0.9,<1.0.0",
]

[project.scripts]
tg-forwarder = "tg_forwarder.cli:main"

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tg_forwarder = [
  "web/static/*",
  "web/static/assets/*",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
